Angular create form model #57

+++ backEnd/ishtar-spring/src/main/java/com/ninos/model/form/Item.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 
 @Data
@@ -26,7 +27,7 @@ public class Item {
     private int quantity;
 
     @Column(name = "price")
-    private int price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "request_order_id")
